pyproject.toml
[project]
name = "mcp-datadog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datadog>=0.51.0",
    "datadog-api-client>=2.33.1",
    "mcp[cli]>=1.4.1",
    "python-dotenv>=1.0.1",
]
